refactor($compile): 메뉴 컨텍스트에 비속어 구현 객체 추가
