+++ src/main/java/kitchenpos/menu/domain/PurgomalumClient.java
@@ -1,0 +1,5 @@
+package kitchenpos.menu.domain;
+
+public interface PurgomalumClient {
+    boolean containsProfanity(String text);
+}
