+++ src/test/java/kitchenpos/menu/application/MenuServiceTest.java
@@ -1,7 +1,7 @@
 package kitchenpos.menu.application;
 
-import kitchenpos.application.FakePurgomalumClient;
-import kitchenpos.products.domain.PurgomalumClient;
+import kitchenpos.menu.infra.FakePurgomalumClient;
+import kitchenpos.menu.domain.PurgomalumClient;
 import kitchenpos.menu.domain.Menu;
 import kitchenpos.menu.domain.MenuGroupRepository;
 import kitchenpos.menu.domain.MenuProduct;
