+++ src/test/java/kitchenpos/menu/infra/FakePurgomalumClient.java
@@ -1,0 +1,19 @@
+package kitchenpos.menu.infra;
+
+import java.util.Arrays;
+import java.util.List;
+import kitchenpos.menu.domain.PurgomalumClient;
+
+public class FakePurgomalumClient implements PurgomalumClient {
+    private static final List<String> profanities;
+
+    static {
+        profanities = Arrays.asList("비속어", "욕설");
+    }
+
+    @Override
+    public boolean containsProfanity(final String text) {
+        return profanities.stream()
+            .anyMatch(profanity -> text.contains(profanity));
+    }
+}
